+++ src/main/core/assigning_descriptors/cpp/Descriptor.h
@@ -12,17 +12,17 @@ using std::map;
 class Descriptor {
 private:
     string name;
-    vector<float> effects;
+    vector<double> effects;
     bool fixed;
     // Fixed descriptors cannot be added or removed from counties after the initital assignments
     // Fixed descriptors CAN have their effects modified
 public:
-    Descriptor(string name, vector<float> effects = {}, bool fixed = false);
+    Descriptor(string name, vector<double> effects = {}, bool fixed = false);
     string getName() const;
-    const vector<float>& getEffects() const;
-    float getEffect(size_t index) const;
-    float setEffect(size_t index, float effect);
-    float addEffect(size_t index, float effect);
+    const vector<double>& getEffects() const;
+    double getEffect(size_t index) const;
+    double setEffect(size_t index, double effect);
+    double addEffect(size_t index, double effect);
     bool isFixed() const;
     bool operator==(const Descriptor& other) const;
     string toString() const;
